Support handler behind proxy.
I tested this bit of code with the 0.0.7 gem and it worked.  This is slightly different however so must be a bit newer.  I modified my code to match this newer branch, but it is untested.  As long as I didn't make any stupid typos I think it should work.

I left the original code in but commented out for now.